+++ src/Provider/Map/Collection.php
@@ -32,16 +32,23 @@ use PSX\Sql\Provider\ProviderCollectionInterface;
 class Collection extends MapAbstract implements ProviderCollectionInterface
 {
     protected $key;
+    protected $filter;
 
-    public function __construct(array $result, array $definition, $key = null)
+    public function __construct(array $result, array $definition, $key = null, \Closure $filter = null)
     {
         parent::__construct($result, $definition);
 
-        $this->key = $key;
+        $this->key    = $key;
+        $this->filter = $filter;
     }
 
     public function getKey()
     {
         return $this->key;
+    }
+
+    public function getFilter()
+    {
+        return $this->filter;
     }
 }